+++ Examples/Write-ProgressEx.autoName.ps1
@@ -5,15 +5,26 @@
 #
 # see https://github.com/mazzy-ax/Write-ProgressEx for details
 
+param(
+    [Parameter(HelpMessage='Specifies how long each iteration sleeps in milliseconds. The default value is used when the user runs this script, 0 is used in unit tests.')]
+    [int]$delayMS = 30
+)
+
+Write-ProgressEx    # reset incomplete runs
+
 $range = 1..500
 
-$range | write-ProgressEx -Total $range.Count | ForEach-Object {
+$range | Write-ProgressEx -Total $range.Count | ForEach-Object {
     # The progress name is equal to the script file name
+    #...
+    Start-Sleep -Milliseconds $delayMS
 }
 
 function test-range {
-    $range | write-ProgressEx -Total $range.Count | ForEach-Object {
+    $range | Write-ProgressEx -Total $range.Count | ForEach-Object {
         # The progress name is equal to the function name
+        #...
+        Start-Sleep -Milliseconds $delayMS
     }
 }