+++ Examples/Write-ProgressEx.pipe.ps1
@@ -5,11 +5,19 @@
 #
 # see https://github.com/mazzy-ax/Write-ProgressEx for details
 
-$nodes = 1..20
-$names = 1..50
+param(
+    [Parameter(HelpMessage='Specifies how long each iteration sleeps in milliseconds. The default value is used when the user runs this script, 0 is used in unit tests.')]
+    [int]$delayMS = 30
+)
 
-$nodes | write-ProgressEx "pipe nodes" -Total $nodes.Count -Status "outer" -ShowMessages -ShowConsoleTitle -ShowElapsed | ForEach-Object {
-    $names | write-ProgressEx "pipe names" -Total $names.Count -id 1 -Status "inner" | ForEach-Object {
+Write-ProgressEx    # reset incomplete runs
+
+$nodes = 1..10
+$names = 1..120
+
+$nodes | Write-ProgressEx "pipe nodes" -Total $nodes.Count -Status "outer" -ShowMessages -ShowConsoleTitle -ShowElapsed | ForEach-Object {
+    $names | Write-ProgressEx "pipe names" -Total $names.Count -id 1 -Status "inner" | ForEach-Object {
         #...
+        Start-Sleep -Milliseconds $delayMS
     }
 }